// src/features/types.ts
/*
    artifact generator: C:\My\wizzi\stfnbssl\wizzi.plugins\packages\wizzi.plugin.ts\lib\artifacts\ts\module\gen\main.js
    package: @wizzi/plugin.ts@
    primary source IttfDocument: C:\My\wizzi\stfnbssl\wizzi.apps\packages\wizzi.hub.backend\.wizzi-override\src\features\types.ts.ittf
    utc time: Fri, 24 Jan 2025 13:13:17 GMT
*/

export type ValidateResult = { 
    isValid: boolean;
    message?: string;
};
;

export type CRUDResult = { 
    oper: string;
    ok: boolean;
    item?: any;
    message?: string;
};
;